resizes dashboard cards, adds hover effects, tweaks background colors
