+++ src/App.css
@@ -12,12 +12,17 @@
     display: flex;
     flex-wrap: wrap;
     gap: 10px;
-    min-height: 60vh;
-    max-width: 70%;
+    padding: 30px;
+    /* min-height: 30vh; */
+    height: 40vh; 
+    overflow-y: auto;
+    /* min-width: 70vw;
+    max-width: 97%; */
+    width: 95%; 
     align-items: center; 
-    justify-content: center;
+    justify-content: ;
     border: 2px solid rgb(255, 255, 255, 0.05);
-    background-color: rgba(212, 212, 212, 0.08);
+    background-color: rgba(212, 212, 212, 0.033);
     border-radius: 20px;
 }
 
@@ -27,18 +32,21 @@
     align-items: center; 
     justify-content: center; 
     color: white;
-    min-width: 20vw;
-    min-height: 20vw;
+    /* max-width: 30vw; */
+    width: calc(100% / 3 - 10px);
+    height: 35vh;
     flex: 0 0 calc(33.33% - 10px);
     border: 2px solid rgb(255, 255, 255, 0.05);
     border-radius: 20px;
     padding: 0.5vw;
-    box-shadow: 2px 2px 15px 0px rgba(255, 255, 255, 0.75);
-    margin: 5vh;
+    box-shadow: 0px 0px 5px 0px rgba(108, 108, 108, 0.75);
+    margin: 1px;
+    transition: all .2s ease-in-out; 
 }
 
 .dashboard-card:hover {
     background-color: rgba(31, 31, 31, 0.134);
-    box-shadow: 2px 2px 25px 0px rgba(255, 255, 255, 0.75);
+    box-shadow: 2px 2px 25px 0px rgba(108, 108, 108, 0.75);
+    transform: scale(1.005);
 }
